+++ changelog.md
@@ -1,3 +1,6 @@
+# v1.0.2
+- Changed Android implementation, it should work without crashing but it won't work properly (without proper testing)
+- Attempted to detect Windows 11 Energy Saver implementation
 # v1.0.1
 - Due to my lack of Android testing, the Android implementation is untested and will break.
 - Removed the Windows Battery Saver implementation. That because Win11 broke it so only works on Win10 for now.